chore: add jest config

+++ src/resolvers/Mutation/deleteTodo.ts
@@ -3,6 +3,10 @@ export default async function deleteTodo(
   { input }: Schema.MutationDeleteTodoArgs,
   context: Context,
 ): Promise<Schema.DeleteTodoPayload> {
+  const existing = await context.repository.get(input.id)
+
+  if (!existing) throw new Error(`Todo not found for id: ${input.id}`)
+
   const todo = await context.repository.remove(input.id)
 
   return { todo }
